Moved some ParticleGroup implementation into cpp file.
